Promotion and Castling UCI support. I added promotion and castling to the UCI, and it don't crash, now, solving some checkmate bugs with search.

+++ src/move.h
@@ -12,7 +12,7 @@ struct Move
 	bool castling;
 	bool mode;
 
-	uint8_t promotion = 255;
+	int promotion = 255;
 	bool capture;
 	uint8_t capturedPiece;
